+++ src/component/_generated/component.ts
@@ -48,6 +48,13 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
         null,
         Name
       >;
+      handleCustomerDeleted: FunctionReference<
+        "mutation",
+        "internal",
+        { stripeCustomerId: string },
+        null,
+        Name
+      >;
       handleCustomerUpdated: FunctionReference<
         "mutation",
         "internal",
@@ -67,6 +74,7 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
           amountDue: number;
           amountPaid: number;
           created: number;
+          metadata?: any;
           status: string;
           stripeCustomerId: string;
           stripeInvoiceId: string;
@@ -124,7 +132,12 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
       handleSubscriptionDeleted: FunctionReference<
         "mutation",
         "internal",
-        { stripeSubscriptionId: string },
+        {
+          cancelAt?: number;
+          cancelAtPeriodEnd?: boolean;
+          currentPeriodEnd?: number;
+          stripeSubscriptionId: string;
+        },
         null,
         Name
       >;
@@ -139,9 +152,22 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
           priceId?: string;
           quantity?: number;
           status: string;
+          stripeCustomerId?: string;
           stripeSubscriptionId: string;
         },
         null,
+        Name
+      >;
+      listSubscriptionsWithCreationTime: FunctionReference<
+        "query",
+        "internal",
+        { stripeCustomerId: string },
+        Array<{
+          _creationTime: number;
+          status: string;
+          stripeCustomerId: string;
+          stripeSubscriptionId: string;
+        }>,
         Name
       >;
       updatePaymentCustomer: FunctionReference<
@@ -172,6 +198,19 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
         string,
         Name
       >;
+      getCheckoutSession: FunctionReference<
+        "query",
+        "internal",
+        { stripeCheckoutSessionId: string },
+        {
+          metadata?: any;
+          mode: string;
+          status: string;
+          stripeCheckoutSessionId: string;
+          stripeCustomerId?: string;
+        } | null,
+        Name
+      >;
       getCustomer: FunctionReference<
         "query",
         "internal",
@@ -181,6 +220,33 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
           metadata?: any;
           name?: string;
           stripeCustomerId: string;
+          userId?: string;
+        } | null,
+        Name
+      >;
+      getCustomerByEmail: FunctionReference<
+        "query",
+        "internal",
+        { email: string },
+        {
+          email?: string;
+          metadata?: any;
+          name?: string;
+          stripeCustomerId: string;
+          userId?: string;
+        } | null,
+        Name
+      >;
+      getCustomerByUserId: FunctionReference<
+        "query",
+        "internal",
+        { userId: string },
+        {
+          email?: string;
+          metadata?: any;
+          name?: string;
+          stripeCustomerId: string;
+          userId?: string;
         } | null,
         Name
       >;
@@ -237,6 +303,19 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
           stripeSubscriptionId: string;
           userId?: string;
         } | null,
+        Name
+      >;
+      listCheckoutSessions: FunctionReference<
+        "query",
+        "internal",
+        { stripeCustomerId: string },
+        Array<{
+          metadata?: any;
+          mode: string;
+          status: string;
+          stripeCheckoutSessionId: string;
+          stripeCustomerId?: string;
+        }>,
         Name
       >;
       listInvoices: FunctionReference<
@@ -360,6 +439,25 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
         }>,
         Name
       >;
+      listSubscriptionsByOrgId: FunctionReference<
+        "query",
+        "internal",
+        { orgId: string },
+        Array<{
+          cancelAt?: number;
+          cancelAtPeriodEnd: boolean;
+          currentPeriodEnd: number;
+          metadata?: any;
+          orgId?: string;
+          priceId: string;
+          quantity?: number;
+          status: string;
+          stripeCustomerId: string;
+          stripeSubscriptionId: string;
+          userId?: string;
+        }>,
+        Name
+      >;
       listSubscriptionsByUserId: FunctionReference<
         "query",
         "internal",
@@ -394,7 +492,7 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
       updateSubscriptionQuantity: FunctionReference<
         "action",
         "internal",
-        { apiKey: string; quantity: number; stripeSubscriptionId: string },
+        { quantity: number; stripeSubscriptionId: string },
         null,
         Name
       >;